Prepare SecurityTestRule for JUnit 5
