+++ java-security-test/src/main/java/com/sap/cloud/security/test/SecurityTestRule.java
@@ -1,65 +1,28 @@
 package com.sap.cloud.security.test;
 
+import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.sap.cloud.security.config.Service;
-import com.sap.cloud.security.test.jetty.JettyTokenAuthenticator;
 import com.sap.cloud.security.token.Token;
-import com.sap.cloud.security.token.TokenClaims;
-import com.sap.cloud.security.token.TokenHeader;
-import com.sap.cloud.security.xsuaa.client.OAuth2ServiceEndpointsProvider;
-import com.sap.cloud.security.xsuaa.client.OAuth2TokenServiceConstants;
-import com.sap.cloud.security.xsuaa.client.XsuaaDefaultEndpoints;
-import org.apache.commons.io.IOUtils;
-import org.eclipse.jetty.annotations.AnnotationConfiguration;
-import org.eclipse.jetty.plus.webapp.EnvConfiguration;
-import org.eclipse.jetty.plus.webapp.PlusConfiguration;
-import org.eclipse.jetty.security.ConstraintSecurityHandler;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.servlet.FilterHolder;
-import org.eclipse.jetty.servlet.ServletHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
-import org.eclipse.jetty.webapp.*;
 import org.junit.rules.ExternalResource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
-import javax.servlet.DispatcherType;
 import javax.servlet.Filter;
 import javax.servlet.Servlet;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.security.NoSuchAlgorithmException;
-import java.security.interfaces.RSAPublicKey;
-import java.security.spec.InvalidKeySpecException;
-import java.util.*;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
-import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
-import static com.sap.cloud.security.xsuaa.client.OidcConfigurationService.DISCOVERY_ENDPOINT_DEFAULT;
 
 public class SecurityTestRule extends ExternalResource {
 
-	public static final String DEFAULT_APP_ID = "xsapp!t0815";
-	public static final String DEFAULT_CLIENT_ID = "sb-clientId!t0815";
-	public static final String DEFAULT_DOMAIN = "localhost";
 	private static final Logger LOGGER = LoggerFactory.getLogger(SecurityTestRule.class);
-	private static final String LOCALHOST_PATTERN = "http://localhost:%d";
-	private final Map<String, ServletHolder> applicationServletsByPath = new HashMap<>();
-	private final List<FilterHolder> applicationServletFilters = new ArrayList<>();
-	// app server
-	private Server applicationServer;
-	private ApplicationServerOptions applicationServerOptions;
-	private boolean useApplicationServer;
 
-	// mock server
-	private WireMockRule wireMockRule;
-	private RSAKeys keys;
-	private int wireMockPort = 0;
-	private Service service;
+	public static final String DEFAULT_APP_ID = SecurityTest.DEFAULT_APP_ID;
+	public static final String DEFAULT_CLIENT_ID = SecurityTest.DEFAULT_CLIENT_ID;
+	public static final String DEFAULT_DOMAIN = SecurityTest.DEFAULT_DOMAIN;
 
-	private String clientId = DEFAULT_CLIENT_ID;
-	private String jwksUrl;
+	SecurityTest base;
 
 	private SecurityTestRule() {
 		// see factory method getInstance()
@@ -74,9 +37,8 @@ public class SecurityTestRule extends ExternalResource {
 	 */
 	public static SecurityTestRule getInstance(Service service) {
 		SecurityTestRule instance = new SecurityTestRule();
-		instance.keys = RSAKeys.generate();
-		instance.service = service;
-		ApplicationServerOptions.forService(service);
+		instance.base = new SecurityTest(service);
+
 		return instance;
 	}
 
@@ -93,7 +55,8 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the rule itself.
 	 */
 	public SecurityTestRule useApplicationServer() {
-		return useApplicationServer(ApplicationServerOptions.forService(service));
+		base.useApplicationServer();
+		return this;
 	}
 
 	/**
@@ -108,8 +71,7 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the rule itself.
 	 */
 	public SecurityTestRule useApplicationServer(ApplicationServerOptions applicationServerOptions) {
-		this.applicationServerOptions = applicationServerOptions;
-		useApplicationServer = true;
+		base.useApplicationServer(applicationServerOptions);
 		return this;
 	}
 
@@ -124,7 +86,7 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the rule itself.
 	 */
 	public SecurityTestRule addApplicationServlet(Class<? extends Servlet> servletClass, String path) {
-		applicationServletsByPath.put(path, new ServletHolder(servletClass));
+		base.addApplicationServlet(servletClass, path);
 		return this;
 	}
 
@@ -139,7 +101,7 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the rule itself.
 	 */
 	public SecurityTestRule addApplicationServlet(ServletHolder servletHolder, String path) {
-		applicationServletsByPath.put(path, servletHolder);
+		base.addApplicationServlet(servletHolder, path);
 		return this;
 	}
 
@@ -152,7 +114,7 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the rule itself.
 	 */
 	public SecurityTestRule addApplicationServletFilter(Class<? extends Filter> filterClass) {
-		applicationServletFilters.add(new FilterHolder(filterClass));
+		base.addApplicationServletFilter(filterClass);
 		return this;
 	}
 
@@ -166,7 +128,7 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the rule itself.
 	 */
 	public SecurityTestRule setPort(int port) {
-		this.wireMockPort = port;
+		base.setPort(port);
 		return this;
 	}
 
@@ -183,22 +145,13 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the rule itself.
 	 */
 	public SecurityTestRule setKeys(String publicKeyPath, String privateKeyPath) {
-		try {
-			this.keys = RSAKeys.fromKeyFiles(publicKeyPath, privateKeyPath);
-		} catch (IOException | NoSuchAlgorithmException | InvalidKeySpecException e) {
-			throw new RuntimeException(e);
-		}
+		base.setKeys(publicKeyPath, privateKeyPath);
 		return this;
 	}
 
 	@Override
 	protected void before() throws Exception {
-		if (useApplicationServer) {
-			startApplicationServer();
-		}
-		setupWireMock();
-
-		// starts WireMock (to stub communication to identity service)
+		base.setup(); // starts WireMock (to stub communication to identity service)
 	}
 
 	/**
@@ -208,21 +161,7 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the preconfigured Jwt token generator
 	 */
 	public JwtGenerator getPreconfiguredJwtGenerator() {
-		JwtGenerator jwtGenerator = JwtGenerator.getInstance(service, clientId)
-				.withPrivateKey(keys.getPrivate());
-		switch (service) {
-		case XSUAA:
-			jwtGenerator
-					.withHeaderParameter(TokenHeader.JWKS_URL, jwksUrl)
-					.withAppId(DEFAULT_APP_ID)
-					.withClaimValue(TokenClaims.XSUAA.GRANT_TYPE, OAuth2TokenServiceConstants.GRANT_TYPE_USER_TOKEN);
-			break;
-		default:
-			jwtGenerator.withClaimValue(TokenClaims.ISSUER, wireMockRule.baseUrl());
-			break;
-		}
-
-		return jwtGenerator;
+		return base.getPreconfiguredJwtGenerator();
 	}
 
 	/**
@@ -233,7 +172,16 @@ public class SecurityTestRule extends ExternalResource {
 	 * @return the token.
 	 */
 	public Token createToken() {
-		return getPreconfiguredJwtGenerator().createToken();
+		return base.createToken();
+	}
+
+	/**
+	 * @deprecated use {@link #getWireMockServer()} method instead.
+	 */
+	@Nullable
+	@Deprecated
+	public WireMockRule getWireMockRule() {
+		throw new UnsupportedOperationException("Deprecated since version 2.6.0. Please use getWireMockServer instead.");
 	}
 
 	/**
@@ -242,112 +190,26 @@ public class SecurityTestRule extends ExternalResource {
 	 * a detailed explanation on how to configure wire mock here:
 	 * http://wiremock.org/docs/getting-started/
 	 *
-	 * @return an instance of WireMockRule
+	 * @return an instance of WireMockRule or null if cast to WireMockRule can not be performed.
 	 */
 	@Nullable
-	public WireMockRule getWireMockRule() {
-		return wireMockRule;
+	public WireMockServer getWireMockServer() {
+		return base.getWireMockServer();
 	}
 
 	/**
 	 * Returns the URI of the embedded jetty server or null if not specified.
-	 * 
+	 *
 	 * @return uri of the application server
 	 */
 	@Nullable
 	public String getApplicationServerUri() {
-		if (useApplicationServer) {
-			return String.format(LOCALHOST_PATTERN, applicationServer.getURI().getPort());
-		}
-		return null;
-	}
-
-	private void setupWireMock() throws IOException {
-		if (wireMockPort == 0) {
-			wireMockRule = new WireMockRule(options().dynamicPort());
-		} else {
-			wireMockRule = new WireMockRule(options().port(wireMockPort));
-		}
-		wireMockRule.start();
-		wireMockPort = wireMockRule.port();
-
-		OAuth2ServiceEndpointsProvider endpointsProvider = new XsuaaDefaultEndpoints(
-				String.format(LOCALHOST_PATTERN, wireMockPort));
-		wireMockRule.stubFor(get(urlEqualTo(endpointsProvider.getJwksUri().getPath()))
-				.willReturn(aResponse().withBody(createDefaultTokenKeyResponse())));
-		wireMockRule.stubFor(get(urlEqualTo(DISCOVERY_ENDPOINT_DEFAULT))
-				.willReturn(aResponse().withBody(createDefaultOidcConfigurationResponse())));
-		jwksUrl = endpointsProvider.getJwksUri().toString();
+		return base.getApplicationServerUri();
 	}
 
 	@Override
 	protected void after() {
-		wireMockRule.shutdown();
-		try {
-			if (useApplicationServer) {
-				applicationServer.stop();
-			}
-		} catch (Exception e) {
-			LOGGER.error("Failed to stop jetty server", e);
-		}
-	}
-
-	private void startApplicationServer() throws Exception {
-		WebAppContext context = createWebAppContext();
-		ServletHandler servletHandler = createServletHandler(context);
-
-		applicationServletsByPath
-				.forEach((path, servletHolder) -> servletHandler.addServletWithMapping(servletHolder, path));
-		applicationServletFilters.forEach((filterHolder) -> servletHandler
-				.addFilterWithMapping(filterHolder, "/*", EnumSet.of(DispatcherType.REQUEST)));
-
-		servletHandler
-				.addFilterWithMapping(new FilterHolder(new SecurityFilter()), "/*", EnumSet.of(DispatcherType.REQUEST));
-
-		applicationServer = new Server(applicationServerOptions.getPort());
-		applicationServer.setHandler(context);
-		applicationServer.start();
-	}
-
-	private ServletHandler createServletHandler(WebAppContext context) {
-		ConstraintSecurityHandler security = new ConstraintSecurityHandler();
-		JettyTokenAuthenticator authenticator = new JettyTokenAuthenticator(
-				applicationServerOptions.getTokenAuthenticator());
-		security.setAuthenticator(authenticator);
-
-		ServletHandler servletHandler = new ServletHandler();
-		security.setHandler(servletHandler);
-		context.setServletHandler(servletHandler);
-		context.setSecurityHandler(security);
-
-		return servletHandler;
-	}
-
-	private WebAppContext createWebAppContext() {
-		WebAppContext context = new WebAppContext();
-		context.setConfigurations(new Configuration[] {
-				new AnnotationConfiguration(), new WebXmlConfiguration(),
-				new WebInfConfiguration(), new PlusConfiguration(), new MetaInfConfiguration(),
-				new FragmentConfiguration(), new EnvConfiguration() });
-		context.setContextPath("/");
-		context.setResourceBase("src/main/java/webapp");
-		context.setParentLoaderPriority(true);
-		return context;
-	}
-
-	private String createDefaultTokenKeyResponse() throws IOException {
-		String encodedPublicKeyModulus = Base64.getUrlEncoder()
-				.encodeToString(((RSAPublicKey) keys.getPublic()).getModulus().toByteArray());
-		String encodedPublicKey = Base64.getEncoder().encodeToString(keys.getPublic().getEncoded());
-		return IOUtils.resourceToString("/token_keys_template.json", StandardCharsets.UTF_8)
-				.replace("$kid", "default-kid")
-				.replace("$public_key", encodedPublicKey)
-				.replace("$modulus", encodedPublicKeyModulus);
-	}
-
-	private String createDefaultOidcConfigurationResponse() throws IOException {
-		return IOUtils.resourceToString("/oidcConfigurationTemplate.json", StandardCharsets.UTF_8)
-				.replace("$issuer", wireMockRule.baseUrl());
+		base.tearDown();
 	}
 
 }
