+++ java-security-test/src/test/java/com/sap/cloud/security/test/SecurityTestRuleTest.java
@@ -92,7 +92,8 @@ public class SecurityTestRuleTest {
 	@Test
 	public void testRuleIsInitializedCorrectly() {
 		assertThat(cut.getApplicationServerUri()).isEqualTo("http://localhost:" + APPLICATION_SERVER_PORT);
-		assertThat(cut.getWireMockRule()).isNotNull();
+		//assertThat(cut.getWireMockRule()).isNotNull();
+		assertThat(cut.getWireMockServer()).isNotNull();
 		assertThat(cut.createToken().getTokenValue())
 				.isEqualTo(cut.getPreconfiguredJwtGenerator().createToken().getTokenValue());
 	}
@@ -148,7 +149,7 @@ public class SecurityTestRuleTest {
 		@Test
 		public void testRuleIsInitializedCorrectly() {
 			assertThat(rule.getApplicationServerUri()).isNull();
-			assertThat(rule.getWireMockRule()).isNotNull();
+			assertThat(rule.getWireMockServer()).isNotNull();
 		}
 	}
